+++ examples/account_management/create_customer.rb
@@ -25,6 +25,7 @@
 require 'optparse'
 require 'google/ads/google_ads'
 require 'date'
+require_relative '../shared/error_handler.rb'
 
 # [START create_customer]
 def create_customer(manager_customer_id)
@@ -91,18 +92,7 @@ if __FILE__ == $0
   begin
     create_customer(options.fetch(:manager_customer_id).tr("-", ""))
   rescue Google::Ads::GoogleAds::Errors::GoogleAdsError => e
-    e.failure.errors.each do |error|
-      STDERR.printf("Error with message: %s\n", error.message)
-      if error.location
-        error.location.field_path_elements.each do |field_path_element|
-          STDERR.printf("\tOn field: %s\n", field_path_element.field_name)
-        end
-      end
-      error.error_code.to_h.each do |k, v|
-        next if v == :UNSPECIFIED
-        STDERR.printf("\tType: %s\n\tCode: %s\n", k, v)
-      end
-    end
-    raise
+    GoogleAdsErrorHandler.handle_google_ads_error(e)
+    raise # Re-raise the error to maintain original script behavior.
   end
 end