+++ examples/basic_operations/add_ad_groups.rb
@@ -20,6 +20,7 @@
 require 'optparse'
 require 'google/ads/google_ads'
 require 'date'
+require_relative '../shared/error_handler.rb'
 
 def add_ad_groups(customer_id, campaign_id)
   # GoogleAdsClient will read a config file from
@@ -86,18 +87,7 @@ if __FILE__ == $0
   begin
     add_ad_groups(options.fetch(:customer_id).tr("-", ""), options[:campaign_id])
   rescue Google::Ads::GoogleAds::Errors::GoogleAdsError => e
-    e.failure.errors.each do |error|
-      STDERR.printf("Error with message: %s\n", error.message)
-      if error.location
-        error.location.field_path_elements.each do |field_path_element|
-          STDERR.printf("\tOn field: %s\n", field_path_element.field_name)
-        end
-      end
-      error.error_code.to_h.each do |k, v|
-        next if v == :UNSPECIFIED
-        STDERR.printf("\tType: %s\n\tCode: %s\n", k, v)
-      end
-    end
-    raise
+    GoogleAdsErrorHandler.handle_google_ads_error(e)
+    raise # Re-raise the error to maintain original script behavior.
   end
 end