+++ examples/planning/generate_keyword_ideas.rb
@@ -19,6 +19,7 @@
 
 require 'optparse'
 require 'google/ads/google_ads'
+require_relative '../shared/error_handler.rb'
 
 # [START generate_keyword_ideas]
 def generate_keyword_ideas(customer_id, location_ids, language_id, keywords,
@@ -103,8 +104,12 @@ if __FILE__ == $0
   # Running the example with -h will print the command line usage.
   options[:customer_id] = 'INSERT_CUSTOMER_ID_HERE'
 
+  # For a list of valid location IDs, see:
+  # https://developers.google.com/google-ads/api/reference/data/geotargets
   options[:location_ids] = ['INSERT_LOCATION_ID_1_HERE', 'INSERT_LOCATION_ID_2_HERE']
 
+  # For a list of valid language IDs, see:
+  # https://developers.google.com/google-ads/api/reference/data/codes-formats#languages
   options[:language_id] = 'INSERT_LANGUAGE_ID_HERE'
 
   # Optional but recommended
@@ -159,18 +164,7 @@ if __FILE__ == $0
       options[:page_url]
     )
   rescue Google::Ads::GoogleAds::Errors::GoogleAdsError => e
-    e.failure.errors.each do |error|
-      STDERR.printf("Error with message: %s\n", error.message)
-      if error.location
-        error.location.field_path_elements.each do |field_path_element|
-          STDERR.printf("\tOn field: %s\n", field_path_element.field_name)
-        end
-      end
-      error.error_code.to_h.each do |k, v|
-        next if v == :UNSPECIFIED
-        STDERR.printf("\tType: %s\n\tCode: %s\n", k, v)
-      end
-    end
-    raise
+    GoogleAdsErrorHandler.handle_google_ads_error(e)
+    raise # Re-raise the error to maintain original script behavior.
   end
 end