+++ examples/remarketing/add_customer_match_user_list.rb
@@ -24,7 +24,7 @@
 # Note: It may take up to several hours for the list to be populated with users.
 # Email addresses must be associated with a Google account.
 # For privacy purposes, the user list size will show as zero until the list has
-# at least 1,000 users. After that, the size will be rounded to the two most
+# at least 100 users. After that, the size will be rounded to the two most
 # significant digits.
 
 require 'optparse'