+++ lib/google/ads/google_ads/v19/services/experiment_service/operations.rb
@@ -135,7 +135,7 @@ module Google
               #   @param options [::Gapic::CallOptions, ::Hash]
               #     Overrides the default settings for this call, e.g, timeout, retries, etc. Optional.
               #
-              # @overload list_operations(name: nil, filter: nil, page_size: nil, page_token: nil)
+              # @overload list_operations(name: nil, filter: nil, page_size: nil, page_token: nil, return_partial_success: nil)
               #   Pass arguments to `list_operations` via keyword arguments. Note that at
               #   least one keyword argument is required. To specify no parameters, or to keep all
               #   the default parameter values, pass an empty Hash as a request object (see above).
@@ -148,6 +148,17 @@ module Google
               #     The standard list page size.
               #   @param page_token [::String]
               #     The standard list page token.
+              #   @param return_partial_success [::Boolean]
+              #     When set to `true`, operations that are reachable are returned as normal,
+              #     and those that are unreachable are returned in the
+              #     [ListOperationsResponse.unreachable] field.
+              #
+              #     This can only be `true` when reading across collections e.g. when `parent`
+              #     is set to `"projects/example/locations/-"`.
+              #
+              #     This field is not by default supported and will result in an
+              #     `UNIMPLEMENTED` error if set unless explicitly documented otherwise in
+              #     service or product specific documentation.
               #
               # @yield [response, operation] Access the result along with the RPC operation
               # @yieldparam response [::Gapic::PagedEnumerable<::Gapic::Operation>]
@@ -711,9 +722,9 @@ module Google
 
                 config_attr :endpoint,      nil, ::String, nil
                 config_attr :credentials,   nil do |value|
-                  allowed = [::String, ::Hash, ::Proc, ::Symbol, ::Google::Auth::Credentials, ::Signet::OAuth2::Client,
-                             nil]
-                  allowed += [::GRPC::Core::Channel, ::GRPC::Core::ChannelCredentials] if defined? ::GRPC
+                  allowed = [::String, ::Hash, ::Proc, ::Symbol, ::Google::Auth::Credentials,
+                             ::Google::Auth::BaseClient, ::Signet::OAuth2::Client, nil]
+                  allowed += [::GRPC::Core::Channel, ::GRPC::Core::ChannelCredentials] if defined? ::GRPC::Core::Channel
                   allowed.any? { |klass| klass === value }
                 end
                 config_attr :scope,         nil, ::String, ::Array, nil